+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 v2.5.0 - [unreleased]
-* Add new pagination classes based on JSON:API recommendations.
+* Add new pagination classes based on JSON:API query parameter *recommendations*:
+  * JsonApiPageNumberPagination and JsonApiLimitOffsetPagination. See [usage docs](docs/usage.md#pagination).
+  * Deprecates PageNumberPagination and LimitOffsetPagination.
 
 v2.4.0 - Released January 25, 2018